Delete the file mockData and created test in case of error
